Remove docker changes from TestJdbcConfig and add them in the singleton class
